Clean-up: use int64 for display ID, instead of int.
The root cause of the bug seems FindDisplayForRootWindow in multi_display_manager.cc,
but here I fixed all wrong "int" declarations.
Verified working on both lumpy/link too.

BUG=144068

Review URL: https://chromiumcodereview.appspot.com/10878005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152783 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d3853ea06d94a0ac90d5fe8abe7e34c163ab15db

+++ ui/aura/display_manager.cc
@@ -28,7 +28,7 @@ bool DisplayManager::use_fullscreen_host_window_ = false;
 
 // static
 gfx::Display DisplayManager::CreateDisplayFromSpec(const std::string& spec) {
-  static int synthesized_display_id = 1000;
+  static int64 synthesized_display_id = 1000;
   gfx::Rect bounds(kDefaultHostWindowX, kDefaultHostWindowY,
                    kDefaultHostWindowWidth, kDefaultHostWindowHeight);
   int x = 0, y = 0, width, height;
